[RENAME] 패키지 수정
issue: #82

* 포인트 내역, 랭킹, 규칙 화면 컴포넌트 파일들을 하나의 패키지에 모아놓도록 변경
