+++ app/src/main/java/com/polzzak_android/presentation/feature/myPage/components/RankingComponents.kt
@@ -1,4 +1,4 @@
-package com.polzzak_android.presentation.feature.myPage.point.ranking
+package com.polzzak_android.presentation.feature.myPage.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -17,7 +17,6 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
